requirements.txt
Click==7.0
feedgen==0.8.0
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.10.3
lxml==4.4.2
MarkupSafe==1.1.1
python-dateutil==2.8.1
PyYAML==5.1.2
six==1.13.0
Werkzeug==0.16.0
